+++ packages/react/src/components/DataTable/stories/DataTable-batch-actions.stories.js
@@ -29,8 +29,11 @@ import DataTable, {
   TableToolbarMenu,
 } from '..';
 
+import { Toggletip, ToggletipButton, ToggletipContent } from '../../Toggletip';
+
 import { batchActionClick, rows, headers } from './shared';
 import mdx from '../DataTable.mdx';
+import Link from '../../Link';
 
 export default {
   title: 'Components/DataTable/Batch Actions',
@@ -282,6 +285,62 @@ export const Playground = (args) => (
         </TableContainer>
       );
     }}
+  </DataTable>
+);
+
+export const Test = () => (
+  <DataTable
+    headers={[
+      {
+        key: 'demo',
+        header: 'Demo',
+      },
+    ]}
+    rows={[{ demo: 'Trigger' }]}>
+    {({
+      rows,
+      headers,
+      getHeaderProps,
+      getRowProps,
+      getTableProps,
+      getTableContainerProps,
+    }) => (
+      <TableContainer {...getTableContainerProps()}>
+        <Table {...getTableProps()}>
+          <TableHead>
+            <TableRow>
+              {headers.map((header) => (
+                <TableHeader key={1} {...getHeaderProps({ header })}>
+                  {header.header}
+                </TableHeader>
+              ))}
+            </TableRow>
+          </TableHead>
+          <TableBody>
+            {rows.map((row) => (
+              <TableRow key={row} {...getRowProps({ row })}>
+                {row.cells.map((cell) => (
+                  <TableCell key={row}>
+                    <Toggletip defaultOpen align="right">
+                      <ToggletipButton>{cell.value}</ToggletipButton>
+                      <ToggletipContent>
+                        <p>
+                          Lorem ipsum dolor{' '}
+                          <Link href="www.google.com" inline>
+                            sit amet
+                          </Link>
+                          .
+                        </p>
+                      </ToggletipContent>
+                    </Toggletip>
+                  </TableCell>
+                ))}
+              </TableRow>
+            ))}
+          </TableBody>
+        </Table>
+      </TableContainer>
+    )}
   </DataTable>
 );
